Move routing utilities from core.helpers to core.routing
