+++ sales/urls.py
@@ -1,7 +1,7 @@
 from rest_framework.urlpatterns import format_suffix_patterns
 from django.urls import path
 
-from core.helpers import merge_sets, gen_url_set
+from core.routing import merge_sets, gen_url_set
 from authentication.views import *
 from .views import *
